Deprecate public access to many mathtext internals.
For most of these classes, there isn't really a "reasonable" way to use
them via public API anyways.

- `Fonts` subclasses are looked up via the private `_font_type_mapping`
  and extending that would require adding new entries for
  `rcParams["mathtext.fontset"]`.
- `FontConstantsBase`\s are looked up via the private
  `_font_constant_mapping`.
- `Node` and its subclasses are clearly internal constructs of the
  mathtext parser (technically, they are visible to `MathtextBackend`
  subclasses when their `render` method is called, but one cannot
  register new `MathtextBackend` subclasses either as that happens via
  the private `_backend_mapping`).
- `Ship` is just an implementation detail for the `ship` callable, which
  can only be called on a `Node` or in a `MathtextBackend`, neither of
  which are really public per the above.
- `Parser` is set up and driven via the (still public) `MathTextParser`.
- `Error`, `SHRINK_FACTOR`, `GROW_FACTOR`, `NUM_SIZE_LEVELS` are helpers
  for the parser.
- `latex_to_*`, `stix_virtual_fonts`, `tex2uni` are helper lookup tables
  for the fonts.  (`tex2uni` stays indirectly accessible via the
  `get_unicode_index`, which basically performs lookup on it.)

Making these private should simplify the implementation of various
improvements to mathtext, e.g. better support for math fonts, as well as
code cleanups.
