+++ .flake8
@@ -50,6 +50,7 @@ per-file-ignores =
 
     lib/matplotlib/__init__.py: F401
     lib/matplotlib/_cm.py: E202, E203, E302
+    lib/matplotlib/_mathtext.py: E221, E251
     lib/matplotlib/_mathtext_data.py: E203, E261
     lib/matplotlib/animation.py: F401
     lib/matplotlib/axes/__init__.py: F401, F403
