+++ doc/api/mathtext_api.rst
@@ -9,3 +9,4 @@
    :members:
    :undoc-members:
    :show-inheritance:
+   :exclude-members: Box, Char, ComputerModernFontConstants, DejaVuSansFontConstants, DejaVuSerifFontConstants, FontConstantsBase, Fonts, Glue, Kern, Node, Parser, STIXFontConstants, STIXSansFontConstants, Ship, StandardPsFonts, TruetypeFonts
