+++ doc/api/next_api_changes/deprecations/18004-AL.rst
@@ -1,4 +1,4 @@
-*facename* parameter of `.mathtext.Fonts.render_glyph`
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+*facename* parameter of ``mathtext.Fonts.render_glyph``
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 This parameter was renamed to *font* for consistency with the rest of the
-`.Fonts` API.
+``Fonts`` API.
